--- a/pkg/tools/list/graph.go
+++ b/pkg/tools/list/graph.go
@@ -2,23 +2,22 @@ package list
 
 import (
 	"fmt"
-	set "github.com/deckarep/golang-set"
 )
 
 // Grapher defines the commonly used interfaces of a graph structure
 type Grapher interface {
 	// AddVertex inserts a node to the graph
-	AddVertex(vertex Vertex) Vertex
-	// DeleteVertex remove a node from the graph by vertex id. The edges that are related to the node will be removed in the meanwhile
-	DeleteVertex(id string) error
+	AddVertex(vertex Vertex)
 	// AddEdge connects two nodes (the source and the target node) in the graph
 	AddEdge(edge Edge)
+	// DeleteVertex remove a node from the graph by vertex id. The edges that are related to the node will be removed in the meanwhile
+	DeleteVertex(id string) error
 	// DeleteEdge will delete the edge which connects a source vertex to the target. The related inbound and outbound edges will be deleted in the meanwhile
 	DeleteEdge(edge Edge) error
 	// ContainsVertex check if a vertex exists in the graph by id
 	ContainsVertex(id string) bool
 	// ContainsEdge checks if there is a directed connection between two nodes in the graph
-	ContainsEdge(sourceId string, targetId string) bool
+	ContainsEdge(source, target Vertex) bool
 	// Vertices returns the list of all the vertices in the graph.
 	Vertices() []Vertex
 	// Edges returns the list of all the edges in the graph.
@@ -32,15 +31,15 @@ type Grapher interface {
 // Graph is a structure that defines a directed graph that contains vertices and directed edges
 type Graph struct {
 	// vertices defines all the vertices in the graph
-	vertices set.Set
+	vertices *IdentifierSet
 	// edges defines all the edges in the graph
-	edges set.Set
+	edges *IdentifierSet
 	// inboundEdges defines a mapping of the vertices and their inbound edges in the graph.
-	// The key is the vertex id and the value is a set of all the inbound edges to the vertex.
-	inboundEdges map[string]set.Set
+	// The key is the vertex id and the value is a map of all the inbound edges to the vertex.
+	inboundEdges map[string]*IdentifierSet
 	// outboundEdges defines a mapping of the vertices and their outbound edges in the graph.
-	// The key is the vertex id and the value is a set of all the outbound edges from the vertex.
-	outboundEdges map[string]set.Set
+	// The key is the vertex id and the value is a map of all the outbound edges from the vertex.
+	outboundEdges map[string]*IdentifierSet
 }
 
 // DirectedAcyclicGraph defines a directed graph with no directed cycles
@@ -55,6 +54,27 @@ func NewDirectedAcyclicGraph() *DirectedAcyclicGraph {
 
 type Identifier interface {
 	Id() string
+}
+
+type IdentifierSet struct {
+	ids map[string]Identifier
+}
+
+func (s *IdentifierSet) Add(id Identifier) {
+	s.ids[id.Id()] = id
+}
+
+func (s *IdentifierSet) Remove(id string) {
+	delete(s.ids, id)
+}
+
+func (s *IdentifierSet) Contains(id string) bool {
+	_, ok := s.ids[id]
+	return ok
+}
+
+func (s *IdentifierSet) Size() int {
+	return len(s.ids)
 }
 
 // Vertex defines the node with in a graph with a unique id for index
@@ -72,16 +92,16 @@ type Edge interface {
 // NewGraph creates an empty graph without any vertices and edges
 func NewGraph() *Graph {
 	return &Graph{
-		vertices:      set.NewSet(),
-		edges:         set.NewSet(),
-		inboundEdges:  make(map[string]set.Set),
-		outboundEdges: make(map[string]set.Set),
+		vertices:      &IdentifierSet{ids: make(map[string]Identifier)},
+		edges:         &IdentifierSet{ids: make(map[string]Identifier)},
+		inboundEdges:  make(map[string]*IdentifierSet),
+		outboundEdges: make(map[string]*IdentifierSet),
 	}
 }
 
 // Size returns the number of the vertices in the graph
 func (g *Graph) Size() int {
-	return g.vertices.Cardinality()
+	return len(g.vertices.ids)
 }
 
 // Draw generates a visual representation of the graph
@@ -90,32 +110,27 @@ func (g *Graph) Draw() string {
 }
 
 // AddVertex inserts a node to the graph
-func (g *Graph) AddVertex(vertex Vertex) Vertex {
-	if vertex == nil {
-		return nil
-	}
+func (g *Graph) AddVertex(vertex Vertex) {
 	g.vertices.Add(vertex)
-	return vertex
 }
 
 // DeleteVertex remove a node from the graph by vertex id. The edges that are related to the node will be removed in the meanwhile
 func (g *Graph) DeleteVertex(id string) error {
-
 	if !g.vertices.Contains(id) {
 		return VertexUnknownError{id}
 	}
 
 	// delete the vertex from its upstream vertices' outbound edges
 	if _, ok := g.inboundEdges[id]; ok {
-		for upstream := range g.inboundEdges[id].Iter() {
-			g.outboundEdges[upstream.(Vertex).Id()].Remove(id)
+		for upstreamId := range g.inboundEdges[id].ids {
+			g.outboundEdges[upstreamId].Remove(id)
 		}
 	}
 
 	// delete the vertex from its downstream vertices' inbound edges
 	if _, ok := g.outboundEdges[id]; ok {
-		for downstream := range g.outboundEdges[id].Iter() {
-			g.inboundEdges[downstream.(Vertex).Id()].Remove(id)
+		for downstreamId := range g.outboundEdges[id].ids {
+			g.inboundEdges[downstreamId].Remove(id)
 		}
 	}
 
@@ -136,17 +151,17 @@ func (g *Graph) ContainsVertex(id string) bool {
 
 // ContainsEdge checks if the edge is contained in the graph
 // The graph contains the edge if there's already an edge which has the same source and target vertex with the given edge
-func (g *Graph) ContainsEdge(source, target string) bool {
-	if _, ok := g.inboundEdges[target]; !ok {
+func (g *Graph) ContainsEdge(source, target Vertex) bool {
+	if _, ok := g.inboundEdges[target.Id()]; !ok {
 		return false
 	}
-	if !g.inboundEdges[target].Contains(source) {
+	if !g.inboundEdges[target.Id()].Contains(source.Id()) {
 		return false
 	}
-	if _, ok := g.outboundEdges[source]; !ok {
+	if _, ok := g.outboundEdges[source.Id()]; !ok {
 		return false
 	}
-	if !g.outboundEdges[source].Contains(target) {
+	if !g.outboundEdges[source.Id()].Contains(target.Id()) {
 		return false
 	}
 	return true
@@ -154,26 +169,26 @@ func (g *Graph) ContainsEdge(source, target string) bool {
 
 // AddEdge adds an edge with the given source and target and records it in the inbound edges and the outbound edges
 func (g *Graph) AddEdge(edge Edge) {
-	if g.ContainsEdge(edge.Source().Id(), edge.Target().Id()) {
+	if g.ContainsEdge(edge.Source(), edge.Target()) {
 		// the graph already contains an edge which connects the edge.Source to the edge.Target
 		return
 	}
 	g.edges.Add(edge)
 
 	if _, ok := g.outboundEdges[edge.Source().Id()]; !ok {
-		g.outboundEdges[edge.Source().Id()] = set.NewSet()
+		g.outboundEdges[edge.Source().Id()] = &IdentifierSet{ids: make(map[string]Identifier)}
 	}
 	g.outboundEdges[edge.Source().Id()].Add(edge.Target())
 
 	if _, ok := g.inboundEdges[edge.Target().Id()]; !ok {
-		g.inboundEdges[edge.Target().Id()] = set.NewSet()
+		g.inboundEdges[edge.Target().Id()] = &IdentifierSet{ids: make(map[string]Identifier)}
 	}
 	g.inboundEdges[edge.Target().Id()].Add(edge.Source())
 }
 
 // DeleteEdge will delete the edge which connects a source vertex to the target. The related inbound and outbound edges will be deleted in the meanwhile
 func (g *Graph) DeleteEdge(edge Edge) error {
-	if !g.ContainsEdge(edge.Source().Id(), edge.Target().Id()) {
+	if !g.ContainsEdge(edge.Source(), edge.Target()) {
 		return EdgeUnknownError{source: edge.Source().Id(), target: edge.Target().Id()}
 	}
 	g.edges.Remove(edge.Id())
@@ -184,7 +199,7 @@ func (g *Graph) DeleteEdge(edge Edge) error {
 
 // visit traverses a DAG from a start node in the specified direction and returns a map of the visited nodes' id
 func visit(g *DirectedAcyclicGraph, id string, visited map[string]bool, down bool) {
-	var nexts set.Set
+	var nexts *IdentifierSet
 	if down {
 		nexts = g.outboundEdges[id]
 	} else {
@@ -192,11 +207,11 @@ func visit(g *DirectedAcyclicGraph, id string, visited map[string]bool, down boo
 	}
 
 	if nexts != nil {
-		for next := range nexts.Iter() {
-			if _, ok := visited[next.(Vertex).Id()]; ok {
+		for next := range nexts.ids {
+			if _, ok := visited[next]; ok {
 				continue
 			}
-			visit(g, next.(Vertex).Id(), visited, down)
+			visit(g, next, visited, down)
 		}
 	}
 
@@ -205,8 +220,8 @@ func visit(g *DirectedAcyclicGraph, id string, visited map[string]bool, down boo
 
 // Vertices returns the list of all the vertices in the graph
 func (g *Graph) Vertices() []Vertex {
-	list := make([]Vertex, 0, g.vertices.Cardinality())
-	for elem := range g.vertices.Iter() {
+	list := make([]Vertex, 0, g.vertices.Size())
+	for _, elem := range g.vertices.ids {
 		list = append(list, elem.(Vertex))
 	}
 	return list
@@ -214,8 +229,8 @@ func (g *Graph) Vertices() []Vertex {
 
 // Edges returns the list of all the edges in the graph
 func (g *Graph) Edges() []Edge {
-	list := make([]Edge, 0, g.edges.Cardinality())
-	for elem := range g.edges.Iter() {
+	list := make([]Edge, 0, g.edges.Size())
+	for _, elem := range g.edges.ids {
 		list = append(list, elem.(Edge))
 	}
 	return list