--- a/pkg/tools/list/testdata/complicate/appops/projectD/base/base.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectD/base/base.k
@@ -1,0 +1,1 @@
+import base.frontend.not_exist