--- a/pkg/tools/list/testdata/complicate/appops/projectE/base/base.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectE/base/base.k
@@ -1,0 +1,2 @@
+import base.frontend.server
+import base.frontend.container