--- a/pkg/tools/list/testdata/complicate/appops/projectF/dev/main.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectF/dev/main.k
@@ -1,0 +1,1 @@
+import ..base.base