--- a/pkg/tools/list/testdata/complicate/appops/projectG/dev/main.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectG/dev/main.k
@@ -1,0 +1,1 @@
+import appops.projectG.base.base