--- a/pkg/tools/list/testdata/complicate/appops/projectH/base/base.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectH/base/base.k
@@ -1,0 +1,3 @@
+# this file exists for testing if a directory can take precedence over a file when they form the same import path
+
+print("this is projectH/base/base.k")