--- a/pkg/tools/list/testdata/complicate/appops/projectH/base/base/base.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectH/base/base/base.k
@@ -1,0 +1,3 @@
+import base.frontend.server.server
+
+print("this is projectH/base/base/base.k")