--- a/pkg/tools/list/testdata/complicate/appops/projectH/dev/main.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectH/dev/main.k
@@ -1,0 +1,1 @@
+import appops.projectH.base.base