--- a/pkg/tools/list/testdata/complicate/appops/projectI/dev/main.k
+++ b/pkg/tools/list/testdata/complicate/appops/projectI/dev/main.k
@@ -1,0 +1,1 @@
+import base.frontend.invalid_a