--- a/pkg/tools/list/testdata/complicate/base/frontend/container/container.k
+++ b/pkg/tools/list/testdata/complicate/base/frontend/container/container.k
@@ -1,3 +1,8 @@
+# short comment
+"""
+doc string
+"""
+
 schema Container:
     name: str = "main"
     command?: [str]