--- a/pkg/tools/list/testdata/complicate/base/frontend/invalid_b/invalid_module_B_1.k
+++ b/pkg/tools/list/testdata/complicate/base/frontend/invalid_b/invalid_module_B_1.k
@@ -1,0 +1,1 @@
+import .invalid_module_B_2