@ldxdl 2022-06-16
1. Clarify your pub APIs clearly, and do not pub unnecessary parts.
2. For pub APIs, pls add docstring about why(purpose, scenes, target-problem...), what(spec, desc, usage..) and how(algorithm, solution..).
3. For important impl structs or functions, pls add comments about core design, algorithm and any necessary desc to help others understand how your code works. You can assume that you are explaining the implementation and examples to others here.
4. Make your core function flow clear and readable, and do abstract by purpose to make your code easy to read and understand.
5. More tests and add benchmark for you pub APIs. Ensure that performance is better than your intended target.