@amyXia1994 2022-07-04
benchmark:

bench test on API: ListDownStreamFiles/ListUpStreamFiles
```
cd ${PROJECT_ROOT}
go test -run=None -bench=BenchmarkImportDepParser_List\.\*

goos: darwin
goarch: amd64
pkg: kusionstack.io/kclvm-go
cpu: Intel(R) Core(TM) i7-7660U CPU @ 2.50GHz
BenchmarkImportDepParser_ListDownStreamFiles-4              1369            752212 ns/op
BenchmarkImportDepParser_ListUpStreamFiles-4                1257            812721 ns/op
PASS
ok      kusionstack.io/kclvm-go 4.515s
```

bench test on walking the importGraph upstream/downstream :
```
cd ${PROJECT_ROOT}/pkg/tools/list
go test -bench=.
goos: darwin
goarch: amd64
pkg: kusionstack.io/kclvm-go/pkg/tools/list
cpu: Intel(R) Core(TM) i7-7660U CPU @ 2.50GHz
BenchmarkImportDepParser_walkDownStream-4         327379              4866 ns/op
BenchmarkImportDepParser_walkUpStreamFiles-4     1666066               676.0 ns/op
PASS
ok      kusionstack.io/kclvm-go/pkg/tools/list  5.663s
```