@amyXia1994 2022-07-04
The PR causes no performance change on API ```ListDownStreamFiles```, and about 37% performance improvement on API ```ListUpStreamFiles```, which is as expected since in this PR the cached value ```processed```(before this PR it's named ```parsed``` field) is checked according to the containing package path at the beginning of the ```func inspect```, but in the older implementation, it was only checked before inspecting ```importPath```(which is not good enough as it missed a situation when the file path ```pkg/b.k``` and its corresponding package path ```pkg/``` are recorded as processed, but the file path ```pkg/a.k``` is not(a.k was processed but not recorded). Then the package path ```pkg``` will be processed multiple times).

I've collected benchmark results five times before and after, then run [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) between old and new:

on API: ListDownStreamFiles/ListUpStreamFiles
```
name                                   old time/op  new time/op  delta
ImportDepParser_ListDownStreamFiles-4  1.13ms ±42%  0.79ms ±17%     ~     (p=0.151 n=5+5)
ImportDepParser_ListUpStreamFiles-4    1.20ms ±17%  0.76ms ±53%  -37.20%  (p=0.032 n=5+5)
```

on walking the importGraph upstream/downstream
```
name                                 old time/op  new time/op  delta
ImportDepParser_walkDownStream-4     4.58µs ±15%  4.20µs ±16%     ~     (p=0.222 n=5+5)
ImportDepParser_walkUpStreamFiles-4  4.51µs ±27%  0.75µs ±10%  -83.32%  (p=0.008 n=5+5)
```